requirements.txt
# PDF智能OCR分类与比对系统 - 依赖包

# ===========================================
# 核心依赖
# ===========================================

# OCR引擎
transformers>=4.35.0     # DeepSeek-OCR2需要
torch>=2.0.0             # PyTorch (CUDA版本)
accelerate>=0.25.0       # 模型加载加速

# PDF处理
pdf2image>=1.16.0        # PDF转图像
PyMuPDF>=1.23.0          # PDF读取和操作 (import fitz)
Pillow>=10.0.0           # 图像处理

# ===========================================
# 可选依赖（备选OCR引擎）
# ===========================================
# paddlepaddle-gpu>=2.5.0  # PaddlePaddle GPU版本
# paddleocr>=2.7.0         # PaddleOCR

# ===========================================
# 中文处理
# ===========================================
jieba>=0.42.1            # 中文分词

# ===========================================
# 数据处理
# ===========================================
numpy>=1.24.0            # 数值计算
opencv-python>=4.8.0     # 图像处理
pandas>=2.0.0            # 数据分析（报告生成）

# ===========================================
# 配置和输出
# ===========================================
pyyaml>=6.0              # YAML配置文件
openpyxl>=3.1.0          # Excel报告生成

# ===========================================
# 进度显示
# ===========================================
tqdm>=4.65.0             # 进度条
rich>=13.0.0             # 终端美化输出
